Split handler tests over several files.
