+++ components/funder/src/handler/tests/mod.rs
@@ -1,0 +1,3 @@
+mod pair_basic;
+mod pair_inconsistency;
+mod utils;
